# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Manager API",
        "description": "REST API for task manager built with Laravel",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/v1/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get tasks list",
                "description": "Returns paginated list of tasks.",
                "operationId": "b07f65b771b24b2f1a573bbaef391dad",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer access token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer 1|abcdef123456789"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Full-text search query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "login"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter tasks by status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "open"
                        }
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "description": "Filter tasks by priority",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "high"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page (max 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of tasks"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Create task",
                "description": "Creates a new task.",
                "operationId": "c5304e96e65e9f60c95ce2e645775883",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer access token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer 1|abcdef123456789"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Task data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "project_id",
                                    "title"
                                ],
                                "properties": {
                                    "project_id": {
                                        "description": "Project ID",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "assignee_id": {
                                        "description": "User assigned to the task",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "title": {
                                        "description": "Task title",
                                        "type": "string",
                                        "example": "Fix OAuth login bug"
                                    },
                                    "description": {
                                        "description": "Task description",
                                        "type": "string",
                                        "example": "Users cannot login via Google OAuth"
                                    },
                                    "priority": {
                                        "description": "Task priority",
                                        "type": "string",
                                        "example": "high"
                                    },
                                    "due_date": {
                                        "description": "Task due date",
                                        "type": "string",
                                        "example": "2026-03-20 12:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task successfully created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "404": {
                        "description": "Project not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/v1/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get task by ID",
                "description": "Returns detailed information about a specific task.",
                "operationId": "642c612eee36cf59133e03508c387380",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer access token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer 1|abcdef123456789"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task details"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Task not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete task",
                "description": "Deletes an existing task.",
                "operationId": "63a569b00eae54d8be881b5ea119cbf4",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer access token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer 1|abcdef123456789"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task deleted"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Task not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Update task",
                "description": "Updates an existing task.",
                "operationId": "f5fda89e7777f13bbb32ff73c4ee5620",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer access token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer 1|abcdef123456789"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Task fields to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Task title",
                                        "type": "string",
                                        "example": "Fix OAuth login bug"
                                    },
                                    "description": {
                                        "description": "Task description",
                                        "type": "string",
                                        "example": "OAuth login fails for some users"
                                    },
                                    "status": {
                                        "description": "Task status",
                                        "type": "string",
                                        "example": "review"
                                    },
                                    "priority": {
                                        "description": "Task priority",
                                        "type": "string",
                                        "example": "high"
                                    },
                                    "assignee_id": {
                                        "description": "User assigned to the task",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "due_date": {
                                        "description": "Task due date",
                                        "type": "string",
                                        "example": "2026-03-20 12:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task successfully updated"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Task not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        }
    },
...
}